Fix for 64-bit OSs.  The Fortran size markers for binary files are longs, not ints.
git-svn-id: file://localhost/data1/git/presto_svn/trunk@1068 5fac236c-a0d8-0310-9745-ef116914eda1

+++ src/barycenter.c
@@ -15,10 +15,10 @@ int read_resid_rec(FILE * file, double *toa, double *obsf)
 /* the file resid2.tmp which is written by TEMPO.       */
 /* It returns 1 if successful, 0 if unsuccessful.       */
 {
-  static int l;
+  static long l;
   static double d[9];
 
-  fread(&l, sizeof(int), 1, file);
+  fread(&l, sizeof(long), 1, file);
   fread(&d, sizeof(double), 9, file);
   /* printf("Barycentric TOA = %17.10f\n", d[0]); */
   /* printf("Postfit residual (pulse phase) = %g\n", d[1]); */
@@ -31,7 +31,7 @@ int read_resid_rec(FILE * file, double *toa, double *obsf)
   /* printf("??? = %g\n\n", d[8]); */
   *toa = d[0];
   *obsf = d[4];
-  return fread(&l, sizeof(int), 1, file);
+  return fread(&l, sizeof(long), 1, file);
 }
 
 void barycenter(double *topotimes, double *barytimes, \
